Create Voice activated qr code application done for PFD
This is project done by my group and I for our PFD module in Ngee Ann polytechnic. It is a voice activated qr code application that aims to solve the problem of hawkers refusing to use qr code payment due to payment discrepancies caused by user input. This app aims to not only solve this problem but also introduce an app that is hawker focused with features such as a breakdown of sales and voice activated qr code to reduce hassle during transaction. This project is still currently being worked on so stay tuned for updates!